Add files via upload
This is the most up-to-date version of the DPEntryEZ Software (originally named `$V6_DPEntryZ_2.0_ROIadjust` - during my initial creation process). This was my third ever created software using Python. Needless to say, during the creation of this software, my primary purpose was to just 'get it to work' and function as I needed it to. By uploading this software to Github, I intend to refactor and reengineer this software so that it can be brought up to par with industry standards.

The purpose of this software is to streamline the process of data entry for drill pipe inspection reports. The software requires a JSON file which contains multiple ROIs containing coordinates, as well as 'keyword' triggers. The user also uploads a scanned PDF document of the drill pipe inspection report (which contains a table containing handwritten data within each cell which each ROI represents) to the software. Once both the PDF document and JSON are uploaded, the user can then begin entering data into the entry widget. Everytime the user presses the Enter key, the next ROI will be displayed. The data is stored in a dictionary until the software closes. This dictionary can then be used to write the data to the official Inspection Report Spreadsheet.